+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - No unreleased changes so far
 
-## [0.37.0] - in testing
+## [0.37.0] - 2022-02-16
 ### Fixed
 - Requests tracking behaviour
 
@@ -14,6 +14,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Possibility to use temperature when startng climatisation
+- Added fail_no_external_power to generic request status
+- Added chargeType attribute to chargingStatus
+- Added chargingSettings attribute to chargingStatus
 
 ## [0.36.4] - 2022-02-12
 ### Fixed